Update __init__.py with correct vertex order ( invert normal) and correct vertex counting
Error happened when there was a mismatch between total vertex count, and the sum of all vertex group vertex count. This was caused by the fact that some vertex were linked to several vertex groups.

Another error was inverting the normals. So i changed the vertex order. It works now.